+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog
 7.1.4 (unreleased)
 ------------------
 
-- If configured with include_textline_fields, enable search and replace on string fields (Archetypes IStringField and dexterity TextLine) 
+- use safe_unicode from CMFPlone
+  [maurits, gotcha]
+
+- If configured with include_textline_fields, enable search and replace on string fields
+  (Archetypes IStringField and dexterity TextLine)
   [gotcha]
 
 - Translate field names in preview table